posting system, appearance, review system

+++ src/main/java/pl/dolien/freetube/controller/AdminController.java
@@ -1,12 +1,14 @@
 package pl.dolien.freetube.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import pl.dolien.freetube.dao.RoleDao;
 import pl.dolien.freetube.entity.Post;
 import pl.dolien.freetube.entity.Review;
 import pl.dolien.freetube.entity.User;
@@ -29,7 +31,7 @@ public class AdminController {
     @Autowired
     private ReviewService reviewService;
 
-    @GetMapping("/")
+    @GetMapping("")
     public String admin(Model m) {
         return users(m);
     }
@@ -42,27 +44,40 @@ public class AdminController {
     }
 
     @GetMapping("/user/delete")
-    public String deleteUser(@RequestParam("theUserName") String userName) {
-        userService.deleteUser(userName);
+    public String deleteUser(@RequestParam("theUserName") String userName, Authentication authentication) {
+        User user = userService.findByUserName(authentication.getName());
+
+        if(user.isAdmin(user)) {
+            userService.deleteUser(userName);
+        }
 
         return "redirect:/admin/users";
     }
 
     @GetMapping("/post/delete")
-    public String deletePost(@RequestParam("postId") int postId, @RequestParam("theUserName") String userName) {
-        postService.deleteById(postId);
+    public String deletePost(@RequestParam("postId") int postId, @RequestParam("theUserName") String userName, Authentication authentication) {
+        User user = userService.findByUserName(authentication.getName());
+
+        if(user.isAdmin(user)) {
+            postService.deleteById(postId);
+        }
+
         return "redirect:/admin/user-posts?theUserName=" + userName;
     }
 
     @GetMapping("/review/delete")
-    public String deleteReview(@RequestParam("reviewId") int reviewId, @RequestParam("theUserName") String userName) {
-        reviewService.deleteById(reviewId);
+    public String deleteReview(@RequestParam("reviewId") int reviewId, @RequestParam("theUserName") String userName, Authentication authentication) {
+        User user = userService.findByUserName(authentication.getName());
+
+        if(user.isAdmin(user)) {
+            reviewService.deleteById(reviewId);
+        }
         return "redirect:/admin/user-reviews?theUserName=" + userName;
     }
 
     @GetMapping("/user-posts")
     public String posts(@RequestParam("theUserName") String userName, Model m) {
-        Iterable<Post> posts = postService.findByUserName(userName);
+        List<Post> posts = postService.findByUserName(userName);
         User user = userService.findByUserName(userName);
         m.addAttribute("user", user);
         m.addAttribute("posts", posts);
