+++ src/main/java/pl/dolien/freetube/controller/ReviewController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import pl.dolien.freetube.dao.RoleDao;
 import pl.dolien.freetube.entity.Post;
 import pl.dolien.freetube.entity.Review;
 import pl.dolien.freetube.entity.User;
@@ -26,7 +27,12 @@ public class ReviewController {
 
     @GetMapping("/delete")
     public String delete(@RequestParam("postId") int postId, @RequestParam("reviewId") int reviewId, Authentication authentication) {
-        reviewService.deleteById(reviewId);
+        String userName = authentication.getName();
+        User user = userService.findByUserName(userName);
+        Review review = reviewService.getById(reviewId);
+        if(review.getUser().equals(user) || user.isAdmin(user)) {
+            reviewService.deleteById(reviewId);
+        }
 
         return "redirect:/post/note?postId=" + postId;
     }
